+++ src/nonim/codegen/expected/zig/statement_variable_immutable_runtime.zig
@@ -1,1 +1,1 @@
-const thing: i64 = 42;
+const thing: isize = 42;
